refactor: 전반적인 계층분리 변경
