+++ BankManager/Sources/BankManager/Customer.swift
@@ -9,10 +9,8 @@ import Foundation
 
 struct Customer {
     private(set) var waitingNumber: Int
-    private(set) var processingTime: Double
     
     init(waitingNumber : Int, processingTime: Double) {
         self.waitingNumber = waitingNumber
-        self.processingTime = processingTime
     }
 }
